+++ crates/genesis_lib/src/setup.rs
@@ -1,5 +1,8 @@
 use bevy::prelude::{Camera2dBundle, Commands, ResMut, SystemSet, Vec2};
 use bevy_rapier2d::prelude::RapierConfiguration;
+use iyes_loopless::prelude::*;
+
+use crate::SimState;
 
 fn camera_setup(mut commands: Commands) {
     commands.spawn(Camera2dBundle::default());
@@ -9,8 +12,10 @@ fn physics_setup(mut rapier_config: ResMut<RapierConfiguration>) {
     rapier_config.gravity = Vec2::ZERO;
 }
 
-pub fn setup_system_set() -> SystemSet {
-    SystemSet::new()
+pub fn sim_setup_system_set() -> SystemSet {
+    ConditionSet::new()
+        .run_in_state(SimState::Simulation)
         .with_system(camera_setup)
         .with_system(physics_setup)
+        .into()
 }
