+++ crates/genesis_lib/src/ui/mod.rs
@@ -15,7 +15,7 @@ pub use interaction::is_paused;
 use iyes_loopless::prelude::*;
 use serde_derive::Serialize;
 
-use crate::statistics;
+use crate::{statistics, SimState};
 
 #[derive(Debug, Serialize)]
 struct RunInfo<'a> {
@@ -103,7 +103,8 @@ pub fn regular_saver(
 }
 
 pub fn interaction_system_set() -> SystemSet {
-    SystemSet::new()
+    ConditionSet::new()
+        .run_in_state(SimState::Simulation)
         .with_system(interaction::move_camera_system)
         .with_system(interaction::camera_zooming_system)
         .with_system(interaction::pause_key_system)
@@ -119,11 +120,13 @@ pub fn interaction_system_set() -> SystemSet {
         .with_system(gui::bug_stats_info_system)
         .with_system(gui::bug_serde_widget)
         .with_system(gui::bug_spawner_widget)
+        .into()
 }
 
 pub fn selection_system_set() -> SystemSet {
     ConditionSet::new()
         .run_if_not(gui::using_ui)
+        .run_in_state(SimState::Simulation)
         .with_system(gui::select_sprite_system)
         .into()
 }
@@ -131,12 +134,14 @@ pub fn selection_system_set() -> SystemSet {
 pub fn manual_spawn_system_set() -> SystemSet {
     ConditionSet::new()
         .run_if_not(gui::using_ui)
+        .run_in_state(SimState::Simulation)
         .with_system(gui::spawn_at_mouse)
         .into()
 }
 
 pub fn global_statistics_system_set() -> SystemSet {
     ConditionSet::new()
+        .run_in_state(SimState::Simulation)
         .run_if_not(interaction::is_paused)
         .with_system(statistics::count_system)
         .with_system(statistics::energy_stats_system)
